Move implementation to `trace_data.rs` file
